fix: update styles.module.css import paths after moving file to airs folder
